# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexusHR VX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityList"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List hired AI employees
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: hireEmployee
      tags: [employees]
      summary: Hire an AI employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HireEmployee"
      responses:
        "201":
          description: Hired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get AI employee details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update AI employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployee"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deactivateEmployee
      tags: [employees]
      summary: Deactivate AI employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /integrations/{toolId}/connect:
    post:
      operationId: connectIntegration
      tags: [integrations]
      summary: Connect an integration
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationItem"
  /integrations/{toolId}/disconnect:
    post:
      operationId: disconnectIntegration
      tags: [integrations]
      summary: Disconnect an integration
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationItem"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [dashboard]
      summary: Analytics overview with charts data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 12
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
...
    Role:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        category:
          type: string
        industry:
          type: string
        seniorityLevel:
          type: string
        description:
          type: string
        coreResponsibilities:
          type: object
        tasks:
          type: object
        toolsAndIntegrations:
          type: object
          nullable: true
        performanceMetrics:
          type: object
          nullable: true
        personalityDefaults:
          type: object
          nullable: true
        skillsTags:
          type: object
          nullable: true
        priceMonthly:
          type: integer
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
      required: [id, title, department, category, industry, seniorityLevel, description, coreResponsibilities, tasks, priceMonthly]
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        roleId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        team:
          type: string
          nullable: true
        status:
          type: string
        personality:
          type: object
          nullable: true
        customInstructions:
          type: string
          nullable: true
        hiredAt:
          type: string
        role:
          $ref: "#/components/schemas/Role"
      required: [id, orgId, roleId, name, status, hiredAt]
    EmployeeList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
    HireEmployee:
      type: object
      properties:
        roleId:
          type: integer
        name:
          type: string
        department:
          type: string
        team:
          type: string
        personality:
          type: object
        customInstructions:
          type: string
      required: [roleId, name]
    UpdateEmployee:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        team:
          type: string
        status:
          type: string
        personality:
          type: object
        customInstructions:
          type: string
...
    IntegrationItem:
      type: object
      properties:
        id:
          type: integer
        toolId:
          type: integer
        status:
          type: string
        connectedAt:
          type: string
          nullable: true
      required: [id, toolId, status]
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
      required: [id, type, title, description, timestamp]
    ActivityList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [data]
    AnalyticsOverview:
      type: object
      properties:
        tasksOverTime:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              completed:
                type: integer
              created:
                type: integer
            required: [date, completed, created]
        utilizationByDepartment:
          type: array
          items:
            type: object
            properties:
              department:
                type: string
              utilization:
                type: number
            required: [department, utilization]
        taskDistribution:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        topPerformers:
          type: array
          items:
            type: object
            properties:
              employeeId:
                type: integer
              name:
                type: string
              tasksCompleted:
                type: integer
              rating:
                type: number
            required: [employeeId, name, tasksCompleted, rating]
      required: [tasksOverTime, utilizationByDepartment, taskDistribution, topPerformers]