# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexusHR VX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoleCoreResponsibilities } from "./roleCoreResponsibilities";
import type { RolePerformanceMetrics } from "./rolePerformanceMetrics";
import type { RolePersonalityDefaults } from "./rolePersonalityDefaults";
import type { RoleSkillsTags } from "./roleSkillsTags";
import type { RoleTasks } from "./roleTasks";
import type { RoleToolsAndIntegrations } from "./roleToolsAndIntegrations";

export interface Role {
  id: number;
  title: string;
  department: string;
  category: string;
  industry: string;
  seniorityLevel: string;
  description: string;
  coreResponsibilities: RoleCoreResponsibilities;
  tasks: RoleTasks;
  toolsAndIntegrations?: RoleToolsAndIntegrations;
  performanceMetrics?: RolePerformanceMetrics;
  personalityDefaults?: RolePersonalityDefaults;
  skillsTags?: RoleSkillsTags;
  priceMonthly: number;
  avatarUrl?: string | null;
  rating?: number | null;
}